+++ src/main/java/chess/dao/BoardDao.java
@@ -1,8 +1,8 @@
 package chess.dao;
 
-import chess.domain.board.service.newDto.BoardModifyRequest;
-import chess.domain.board.service.newDto.BoardRegisterRequest;
-import chess.domain.board.service.newDto.BoardSearchResponse;
+import chess.domain.board.service.dto.BoardModifyRequest;
+import chess.domain.board.service.dto.BoardRegisterRequest;
+import chess.domain.board.service.dto.BoardSearchResponse;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
