+++ src/test/java/chess/domain/board/service/mapper/BoardMapperTest.java
@@ -3,7 +3,7 @@ package chess.domain.board.service.mapper;
 import chess.domain.board.Board;
 import chess.domain.board.Turn;
 import chess.domain.board.position.Position;
-import chess.domain.board.service.newDto.BoardRegisterRequest;
+import chess.domain.board.service.dto.BoardRegisterRequest;
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.jumper.King;
